Create conditionalOperator
Conditional operators come in handy when you need to assign a certain value to a single variable. A condition will determine what the value is, but that information is going to end up in the same variable no matter what. In this case, depending on the amount of miles driven, you are going to get paid a certain dollar amount. However, since there are other consideration taken into account if you drive over 100 miles, you will use a conditional operator. You could've used an if/else statement.